Refactored image build to generate database from within the docker build to resolve missing media files
As a result of 404'ing images in pub/media it was discovered that (for example) the setup:install routine generates new images in pub/media for configurable products. For example, the image wsh03-gray_main.jpg is installed by the module, but after setup:install wsh03-gray_main_1.jpg and wsh03-gray_main_2.jpg are also present, so separating the build routines for these caused the used images to not be found. Combining them like this should result in images all being present no matter if they are installed during setup:install or on module deployment.

This also addresses the need to previously copy in styles.css fixtures to pub/media/styles.css which is also created by the setup:install process.